+++ src/main/java/com/iliasdev/dto/ExchangeCurrencyDto.java
@@ -4,12 +4,14 @@ import com.iliasdev.model.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class ExchangeCurrencyDto {
     private CurrencyModel baseCurrencyModel;
     private CurrencyModel targetCurrencyModel;
-    private double rate;
-    private double amount;
-    private double convertedAmount;
+    private BigDecimal rate;
+    private BigDecimal amount;
+    private BigDecimal convertedAmount;
 }
